# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SecuritySuite (secapi) — APIs propias",
    "version": "0.1.0",
    "description": "Catálogo generado por `pnpm docs:api` a partir de `src/app/api/**/route.ts`.\n\nLo generado describe **lo que el código hace hoy**, incluida la autenticación real de\ncada handler. Las descripciones a mano viven en `docs/api/anotaciones.yaml` y ganan\nsobre lo inferido; `GET /api/docs/spec` devuelve el documento ya mergeado.\n\nNo editar este archivo a mano: se regenera."
  },
  "servers": [
    {
      "url": "http://localhost:4005",
      "description": "desarrollo local (pnpm dev)"
    },
    {
      "url": "https://secapi.riogas.com.uy",
      "description": "secapi (PM2, puerto 3001)"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerJwt": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT emitido por `POST /api/db/login`. Payload: `{ iss, username, userId, sistema }` — no lleva el flag root."
      },
      "cookieToken": {
        "type": "apiKey",
        "in": "cookie",
        "name": "token",
        "description": "El mismo JWT, tomado de la cookie `token` cuando no hay header."
      },
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    }
  },
  "paths": {
...
    "/api/db/admin/import-sgm-preferences": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Handler principal",
        "description": "Handler principal",
        "operationId": "post_api_db_admin_import_sgm_preferences",
        "parameters": [
          {
            "name": "dryRun",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Detectado estáticamente (`searchParams.get`); anotalo para describirlo."
          }
        ],
        "responses": {
          "default": {
            "description": "Sin anotar todavía (ver docs/api/README.md)."
          }
        },
        "x-auth": {
          "modo": "jwt",
          "verificaFirma": true,
          "detectadoEn": "handler",
          "detalle": "JWT en `Authorization: Bearer` o cookie `token`, con firma y vencimiento verificados por el guard central de `/api/db` (`requireApiAuth`, src/lib/auth/apiGuard.ts). El nivel exigido por esta ruta está declarado en `POLITICAS`, en ese mismo archivo."
        },
        "x-archivo": "src/app/api/db/admin/import-sgm-preferences/route.ts",
        "x-generado": true,
        "security": [
          {
            "bearerJwt": []
          },
          {
            "cookieToken": []
          }
        ]
      }
    },
...
    "/api/db/aplicaciones/{id}/roles": {
      "get": {
        "tags": [
          "aplicaciones"
        ],
        "summary": "Query params: estado (A | I | \"\" para todos)",
        "description": "GET /api/db/aplicaciones/[id]/roles\nQuery params: estado (A | I | \"\" para todos)",
        "operationId": "get_api_db_aplicaciones_id_roles",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Segmento dinámico `[id]` de la ruta."
          },
          {
            "name": "estado",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Detectado estáticamente (`searchParams.get`); anotalo para describirlo."
          }
        ],
        "responses": {
          "default": {
            "description": "Sin anotar todavía (ver docs/api/README.md)."
          }
        },
        "x-auth": {
          "modo": "jwt",
          "verificaFirma": true,
          "detectadoEn": "handler",
          "detalle": "JWT en `Authorization: Bearer` o cookie `token`, con firma y vencimiento verificados por el guard central de `/api/db` (`requireApiAuth`, src/lib/auth/apiGuard.ts). El nivel exigido por esta ruta está declarado en `POLITICAS`, en ese mismo archivo."
        },
        "x-archivo": "src/app/api/db/aplicaciones/[id]/roles/route.ts",
        "x-generado": true,
        "security": [
          {
            "bearerJwt": []
          },
          {
            "cookieToken": []
          }
        ]
      }
    },
...
  },
...
}